fix(guardian): resolve all Guardian violations for sprint-4
- Barrel exports: add useAgentSettings, useUpdateAgentSettings, AppBehaviorSection to settings/index.ts
- Barrel exports: add all hooks, gitOverviewKeys, and new components to git-overview/index.ts
- Size: extract FilesChangedTab + ErrorsTab to AgentPanelTabs.tsx (AgentPanelExpanded: 320→215 lines)
- Size: extract CreatePrDialog to separate file (CommitPanel: 312→186 lines)
- Design system: replace raw <p> with <Text> in AgentPanelCompact, AgentPanelExpanded, AgentLayoutGrid, AgentLayoutSingle
- Design system: replace raw <h2> with <Heading> in AppBehaviorSection
- Design system: replace && with ternary in AgentPanelExpanded, AgentLayoutSingle, GitStatusCard
- AgentDashboardPage (603 lines) was already 518 lines pre-sprint-4 — flagged, not refactored
