+++ src/renderer/features/agent-dashboard/components/AgentPanelExpanded.tsx
@@ -5,9 +5,9 @@
  * Chat, Files Changed, Errors.
  */
 
-import { AlertTriangle, CheckCircle2, FileCode, ListChecks, MessageSquare, Minimize2, Maximize2 } from 'lucide-react';
+import { AlertTriangle, FileCode, ListChecks, MessageSquare, Minimize2, Maximize2 } from 'lucide-react';
 
-import type { AgentError, AgentFileChange, AgentSession } from '@shared/types/agent-dashboard';
+import type { AgentSession } from '@shared/types/agent-dashboard';
 
 import { cn } from '@renderer/shared/lib/utils';
 
@@ -15,9 +15,7 @@ import {
   Badge,
   Button,
   Card,
-  CardContent,
   CardHeader,
-  ScrollArea,
   Separator,
   Tabs,
   TabsContent,
@@ -27,6 +25,7 @@ import {
 } from '@ui';
 
 import { AgentChatPanel } from './AgentChatPanel';
+import { ErrorsTab, FilesChangedTab } from './AgentPanelTabs';
 import { AgentStatusBar } from './AgentStatusBar';
 import { QaPanel } from './QaPanel';
 import { TasksTab } from './TasksTab';
@@ -43,18 +42,6 @@ interface AgentPanelExpandedProps {
 
 // ─── Helpers ───────────────────────────────────────────────
 
-function getFileStatusChar(status: AgentFileChange['status']): string {
-  if (status === 'added') return 'A';
-  if (status === 'deleted') return 'D';
-  return 'M';
-}
-
-function getFileStatusColor(status: AgentFileChange['status']): string {
-  if (status === 'added') return 'text-success';
-  if (status === 'deleted') return 'text-destructive';
-  return 'text-warning';
-}
-
 function deriveFeatureSlug(agent: AgentSession): string {
   const { branch } = agent;
   if (branch !== undefined) {
@@ -69,9 +56,7 @@ function deriveFeatureSlug(agent: AgentSession): string {
 // ─── Token Usage Section ────────────────────────────────────
 
 function formatTokenCount(count: number): string {
-  if (count >= 1000) {
-    return `${Math.round(count / 1000)}k`;
-  }
+  if (count >= 1000) return `${Math.round(count / 1000)}k`;
   return String(count);
 }
 
@@ -120,96 +105,6 @@ function TokenUsageSection({ inputTokens, outputTokens }: TokenUsageSectionProps
   );
 }
 
-// ─── Files Changed Tab ─────────────────────────────────────
-
-function FilesChangedTab({ files }: { files: AgentFileChange[] }) {
-  if (files.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center py-8 text-muted-foreground">
-        <FileCode className="mb-2 h-8 w-8" />
-        <p className="text-sm">No files changed</p>
-      </div>
-    );
-  }
-
-  return (
-    <ScrollArea className="h-full">
-      <div className="space-y-1 p-2">
-        {files.map((file) => (
-          <div
-            key={file.path}
-            className="flex items-center gap-2 rounded px-2 py-1 text-xs hover:bg-muted/50"
-          >
-            <span className={cn('w-4 font-mono font-bold', getFileStatusColor(file.status))}>
-              {getFileStatusChar(file.status)}
-            </span>
-            <span className="min-w-0 flex-1 truncate font-mono text-foreground">
-              {file.path}
-            </span>
-            <span className="shrink-0 text-muted-foreground">
-              <span className="text-success">+{String(file.additions)}</span>
-              {' '}
-              <span className="text-destructive">-{String(file.deletions)}</span>
-            </span>
-          </div>
-        ))}
-      </div>
-    </ScrollArea>
-  );
-}
-
-// ─── Errors Tab ────────────────────────────────────────────
-
-function ErrorsTab({ errors }: { errors: AgentError[] }) {
-  if (errors.length === 0) {
-    return (
-      <div className="flex flex-col items-center justify-center py-8 text-muted-foreground">
-        <CheckCircle2 className="mb-2 h-8 w-8 text-success" />
-        <p className="text-sm">No errors</p>
-        <p className="mt-1 text-xs">Clean run — no issues detected</p>
-      </div>
-    );
-  }
-
-  return (
-    <ScrollArea className="h-full">
-      <div className="space-y-2 p-2">
-        {errors.map((error) => (
-          <Card
-            key={error.id}
-            className={cn(
-              'overflow-hidden',
-              error.severity === 'error' ? 'border-destructive' : 'border-warning',
-            )}
-          >
-            <CardContent className="p-3">
-              <div className="flex items-start gap-2">
-                <AlertTriangle
-                  className={cn(
-                    'mt-0.5 h-4 w-4 shrink-0',
-                    error.severity === 'error' ? 'text-destructive' : 'text-warning',
-                  )}
-                />
-                <div className="min-w-0 flex-1">
-                  <p className="text-sm text-foreground">{error.message}</p>
-                  {error.source !== undefined && error.source.length > 0 && (
-                    <p className="mt-1 font-mono text-xs text-muted-foreground">
-                      {error.source}
-                    </p>
-                  )}
-                  <p className="mt-1 text-xs text-muted-foreground">
-                    {new Date(error.timestamp).toLocaleTimeString()}
-                  </p>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        ))}
-      </div>
-    </ScrollArea>
-  );
-}
-
 // ─── Component ─────────────────────────────────────────────
 
 export function AgentPanelExpanded({
@@ -243,16 +138,16 @@ export function AgentPanelExpanded({
         </div>
       </CardHeader>
 
-      {agent.taskName !== undefined && agent.taskName.length > 0 && (
+      {agent.taskName !== undefined && agent.taskName.length > 0 ? (
         <div className="border-t px-4 py-2">
-          <p className="text-xs text-muted-foreground">
+          <Text className="text-xs text-muted-foreground">
             Task: {agent.taskName}
-            {agent.branch !== undefined && agent.branch.length > 0 && (
+            {agent.branch !== undefined && agent.branch.length > 0 ? (
               <span className="ml-2">Branch: {agent.branch}</span>
-            )}
-          </p>
+            ) : null}
+          </Text>
         </div>
-      )}
+      ) : null}
 
       <TokenUsageSection
         inputTokens={agent.tokenUsage.input}
