+++ src/renderer/features/git-overview/index.ts
@@ -3,3 +3,21 @@
  */
 
 export { GitPage } from './components/GitPage';
+export { GitStatusCard } from './components/GitStatusCard';
+export { BranchList } from './components/BranchList';
+export { WorktreeList } from './components/WorktreeList';
+export { CommitPanel } from './components/CommitPanel';
+
+export {
+  useGitStatus,
+  useGitBranches,
+  useListWorktrees,
+  useCreateBranch,
+  useCreateWorktree,
+  useRemoveWorktree,
+  useGitCommit,
+  useGitPush,
+  useCreatePr,
+  useGitOverviewEvents,
+} from './api/useGit';
+export { gitOverviewKeys } from './api/queryKeys';
